Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

BundleAdjustmentModels.BundleAdjustmentModel

BundleAdjustmentModels.BundleAdjustmentModel

BundleAdjustmentModels.delete_all_ba_artifacts!

BundleAdjustmentModels.delete_ba_artifact!

BundleAdjustmentModels.fetch_ba_group

BundleAdjustmentModels.fetch_ba_name

BundleAdjustmentModels.problems_df

​

BundleAdjustmentModels.BundleAdjustmentModel — Type

Represent a bundle adjustement problem in the form

minimize   ½ ‖F(x)‖²

where F(x) is the vector of residuals.

BundleAdjustmentModels.BundleAdjustmentModel — Method

BundleAdjustmentModel(name::AbstractString; T::Type=Float64)

Constructor of BundleAdjustmentModel, creates an NLSModel with name name from a BundleAdjustment archive with precision T.

BundleAdjustmentModels.delete_all_ba_artifacts! — Method

delete_all_ba_artifacts!()

Delete all the BundleAdjustmentModel artifacts from the artifact store.

BundleAdjustmentModels.delete_ba_artifact! — Method

delete_ba_artifact!(name::AbstractString)

Delete the artifact name from the artifact store.

BundleAdjustmentModels.fetch_ba_group — Method

fetch_ba_group(group::AbstractString)

Get all the problems with the group name group. Return an array of the paths where the problems are stored. Group possibilities are : "trafalgar", "venice", "dubrovnik" and "ladybug".

BundleAdjustmentModels.fetch_ba_name — Method

fetch_ba_name(name::AbstractString)

Get the problem with name name. Return the path where the problem is stored.

BundleAdjustmentModels.problems_df — Method

problems_df()

Return a dataframe of the problems and their characteristics.