Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

BALNLSModels.BALNLSModel

BALNLSModels.BALNLSModel

BALNLSModels.BALNLSModel

BALNLSModels.delete_all_balartifacts!

BALNLSModels.delete_balartifact!

BALNLSModels.fetch_bal_group

BALNLSModels.fetch_bal_name

BALNLSModels.get_first_name_and_group

BALNLSModels.problems_df

​

BALNLSModels.BALNLSModel — Type

Represent a bundle adjustement problem in the form

minimize    0
subject to  F(x) = 0,

where F(x) is the vector of residuals.

BALNLSModels.BALNLSModel — Method

BALNLSModel(name::AbstractString, group::AbstractString, T::Type=Float64)

Alternate constructor of BALNLSModel Get the path of the problem name name from the group group with the precision T. Return a NLSModel generated from this problem data using NLPModels

BALNLSModels.BALNLSModel — Method

BALNLSModel(filename::AbstractString; T::Type=Float64, verbose::Bool=false)

Constructor of BALNLSModel, creates an NLSModel from a BAL archive

BALNLSModels.delete_all_balartifacts! — Method

delete_all_balartifacts!()

Delete all the BALNLSModels artifacts from the artifact store

BALNLSModels.delete_balartifact! — Method

delete_balartifact!(name::AbstractString, group::AbstractString)

Delete the artifact name from the artifact store

BALNLSModels.fetch_bal_group — Method

fetch_bal_group(group::AbstractString)

Get all the problems with the group name group. Return an array of the paths where the problems are stored. Group possibilities are : trafalgar, venice, dubrovnik and ladybug

BALNLSModels.fetch_bal_name — Method

fetch_bal_name(name::AbstractString, group::AbstractString)

Get the problem with name name from the group group. Return the path where the problem is stored.

BALNLSModels.get_first_name_and_group — Method

get_first_name_and_group(dataframe::DataFrame)

Return the name and the group of the first row in the DataFrame

BALNLSModels.problems_df — Method

problems_df()

Return a dataframe of the problems and their characteristics