Reference

​

Contents

​

Reference

Contents

Index

​

Index

​

BundleAdjustmentModels.BundleAdjustmentModel

BundleAdjustmentModels.BundleAdjustmentModel

BundleAdjustmentModels.BundleAdjustmentModel

BundleAdjustmentModels.delete_all_ba_artifacts!

BundleAdjustmentModels.delete_ba_artifact!

BundleAdjustmentModels.fetch_ba_group

BundleAdjustmentModels.fetch_ba_name

BundleAdjustmentModels.get_first_name_and_group

BundleAdjustmentModels.problems_df

​

BundleAdjustmentModels.BundleAdjustmentModel — Type

Represent a bundle adjustement problem in the form

minimize    0
subject to  F(x) = 0,

where F(x) is the vector of residuals.

BundleAdjustmentModels.BundleAdjustmentModel — Method

BundleAdjustmentModel(name::AbstractString, group::AbstractString, T::Type=Float64)

Alternate constructor of BundleAdjustmentModel Get the path of the problem name name from the group group with the precision T. Return a NLSModel generated from this problem data using NLPModels

BundleAdjustmentModels.BundleAdjustmentModel — Method

BundleAdjustmentModel(filename::AbstractString; T::Type=Float64, verbose::Bool=false)

Constructor of BundleAdjustmentModel, creates an NLSModel from a BundleAdjustment archive

BundleAdjustmentModels.delete_all_ba_artifacts! — Method

delete_all_ba_artifacts!()

Delete all the BundleAdjustmentModels artifacts from the artifact store

BundleAdjustmentModels.delete_ba_artifact! — Method

delete_ba_artifact!(name::AbstractString, group::AbstractString)

Delete the artifact name from the artifact store

BundleAdjustmentModels.fetch_ba_group — Method

fetch_ba_group(group::AbstractString)

Get all the problems with the group name group. Return an array of the paths where the problems are stored. Group possibilities are : trafalgar, venice, dubrovnik and ladybug

BundleAdjustmentModels.fetch_ba_name — Method

fetch_ba_name(name::AbstractString, group::AbstractString)

Get the problem with name name from the group group. Return the path where the problem is stored.

BundleAdjustmentModels.get_first_name_and_group — Method

get_first_name_and_group(dataframe::DataFrame)

Return the name and the group of the first row in the DataFrame

BundleAdjustmentModels.problems_df — Method

problems_df()

Return a dataframe of the problems and their characteristics